[chore] #99 클래스명 수정
